# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Project NETRA API
  version: 1.0.0
  description: |
    Networked Eyes for Tactical Response and Awareness.
    Citizen-Consented, Camera-Agnostic, Hybrid Video-Intelligence Platform.
    
    **Governance rules enforced at API layer:**
    - All investigation-scoped operations require valid `investigation_id`
    - Face recognition requires officer attestation for every match
    - Re-ID is investigation-bound — global archive search has no API path
    - Watchlist operations require SENIOR_SP role
    - All operations are audit-logged
servers:
  - url: https://api.netra.gov.in/v1
    description: Production (India sovereign cloud)
  - url: http://localhost:3000/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /auth/citizen/otp/send:
    post:
      tags: [Auth]
      summary: Send OTP to citizen phone
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  pattern: '^\+91\d{10}$'
                  example: '+919876543210'
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: OTP sent
        '429':
          description: Too many requests (3 per minute limit)
  /auth/citizen/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and authenticate citizen
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, otp]
              properties:
                phone:
                  type: string
                  example: '+919876543210'
                otp:
                  type: string
                  minLength: 6
                  maxLength: 6
                  example: '123456'
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Invalid or expired OTP
  /auth/officer/login:
    post:
      tags: [Auth]
      summary: Officer login (username + password)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens issued or MFA required
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthTokens'
                  - $ref: '#/components/schemas/MfaRequired'
  /auth/officer/mfa/verify:
    post:
      tags: [Auth]
      summary: Complete MFA verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfaToken, totp]
              properties:
                mfaToken:
                  type: string
                totp:
                  type: string
                  minLength: 6
                  maxLength: 6
      responses:
        '200':
          description: Authentication complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'

  # ── Cameras ─────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        isNewUser:
          type: boolean
    MfaRequired:
      type: object
      properties:
        requiresMfa:
          type: boolean
          example: true
        mfaToken:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          minLength: 8
...